Avoid rerendering across the stack
