+++ src/components/ChatDisplay.tsx
@@ -1,12 +1,33 @@
-import React, { useState, useEffect, useMemo } from "react";
-import { Box, Text, useInput, useStdout } from "ink";
-import TextInput from "ink-text-input";
+import React, { useState, useEffect, useMemo, memo } from "react";
+import { Box, Text, useStdout } from "ink";
 import Spinner from "ink-spinner";
-import { ChatMessage, MentionableUser } from "../models/chat.js";
+import { ChatMessage } from "../models/chat.js";
 import { IssueCommand } from "../issues/types.js";
-import { ChangeReviewer, fetchEligibleReviewers } from "../lib/clients/baz.js";
-import MentionAutocomplete from "./MentionAutocomplete.js";
 import { renderMarkdown } from "../lib/markdown.js";
+import ChatInput from "./ChatInput.js";
+
+const MemoizedMessage = memo<{ message: ChatMessage }>(
+  ({ message }) => {
+    const renderedContent = useMemo(
+      () => renderMarkdown(message.content),
+      [message.content],
+    );
+
+    return (
+      <Box flexDirection="column" marginBottom={1}>
+        <Text bold color={message.role === "user" ? "cyan" : "yellow"}>
+          {message.role === "user" ? "You:" : "Baz:"}
+        </Text>
+        <Box paddingLeft={2}>
+          <Text>{renderedContent}</Text>
+        </Box>
+      </Box>
+    );
+  },
+  (prevProps, nextProps) =>
+    prevProps.message.content === nextProps.message.content &&
+    prevProps.message.role === nextProps.message.role,
+);
 
 interface ChatDisplayProps {
   messages: ChatMessage[];
@@ -20,271 +41,100 @@ interface ChatDisplayProps {
   onBack: () => void;
 }
 
-const ChatDisplay: React.FC<ChatDisplayProps> = ({
-  messages,
-  isLoading,
-  onSubmit,
-  placeholder = "How will it affect the code?",
-  availableCommands = [],
-  disabled = false,
-  prId,
-  enableMentions = false,
-  onBack,
-}) => {
-  const { stdout } = useStdout();
-  const [inputValue, setInputValue] = useState("");
-  const [showFullHelp, setShowFullHelp] = useState(false);
-  const [reviewers, setReviewers] = useState<ChangeReviewer[]>([]);
-  const [showMentionAutocomplete, setShowMentionAutocomplete] = useState(false);
-  const [mentionSearchQuery, setMentionSearchQuery] = useState("");
-  const [mentionStartIndex, setMentionStartIndex] = useState(-1);
-  const [inputKey, setInputKey] = useState(0);
-  const [terminalWidth, setTerminalWidth] = useState(stdout?.columns || 80);
+const ChatDisplay = memo<ChatDisplayProps>(
+  ({
+    messages,
+    isLoading,
+    onSubmit,
+    placeholder = "How will it affect the code?",
+    availableCommands = [],
+    disabled = false,
+    prId,
+    enableMentions = false,
+    onBack,
+  }) => {
+    const { stdout } = useStdout();
+    const [terminalWidth, setTerminalWidth] = useState(stdout?.columns || 80);
 
-  useEffect(() => {
-    const handleResize = () => {
-      if (stdout?.columns) {
-        setTerminalWidth(stdout.columns);
-        setInputKey((prev) => prev + 1);
-      }
-    };
-
-    if (stdout) {
-      stdout.on("resize", handleResize);
-
-      return () => {
-        stdout.off("resize", handleResize);
-      };
-    }
-  }, [stdout]);
-
-  useEffect(() => {
-    if (enableMentions && prId) {
-      fetchEligibleReviewers(prId)
-        .then(setReviewers)
-        .catch((error) => {
-          console.error("Failed to fetch eligible reviewers:", error);
-        });
-    }
-  }, [enableMentions, prId]);
-
-  useInput(
-    (input, key) => {
-      if (key.escape && !isLoading && !disabled) {
-        if (showMentionAutocomplete) {
-          setShowMentionAutocomplete(false);
-          setMentionSearchQuery("");
-          setMentionStartIndex(-1);
-        } else {
-          onBack();
+    useEffect(() => {
+      const handleResize = () => {
+        if (stdout?.columns) {
+          setTerminalWidth(stdout.columns);
         }
+      };
+
+      if (stdout) {
+        stdout.on("resize", handleResize);
+
+        return () => {
+          stdout.off("resize", handleResize);
+        };
       }
-      if (input === "?" && !isLoading && !disabled && inputValue === "") {
-        setShowFullHelp((prev) => !prev);
-      }
-    },
-    { isActive: !showMentionAutocomplete },
-  );
+    }, [stdout]);
 
-  const handleInputChange = (value: string) => {
-    if (inputValue === "" && value.startsWith("?")) {
-      setInputKey((prev) => prev + 1);
-      return;
-    }
-
-    setInputValue(value);
-
-    if (value.startsWith("/") && !value.includes(" ")) {
-      setShowFullHelp(true);
-    } else if (showFullHelp && !value.startsWith("/")) {
-      setShowFullHelp(false);
-    }
-
-    if (!enableMentions) {
-      return;
-    }
-
-    const lastAtIndex = value.lastIndexOf("@");
-    if (lastAtIndex !== -1) {
-      const textAfterAt = value.slice(lastAtIndex + 1);
-      if (!textAfterAt.includes(" ")) {
-        setShowMentionAutocomplete(true);
-        setMentionSearchQuery(textAfterAt);
-        setMentionStartIndex(lastAtIndex);
-      } else {
-        setShowMentionAutocomplete(false);
-        setMentionSearchQuery("");
-        setMentionStartIndex(-1);
-      }
-    } else {
-      setShowMentionAutocomplete(false);
-      setMentionSearchQuery("");
-      setMentionStartIndex(-1);
-    }
-  };
-
-  const handleMentionSelect = (reviewer: MentionableUser) => {
-    const lastSlashIndex = reviewer.login.lastIndexOf("/");
-    const login = reviewer.login.substring(lastSlashIndex + 1);
-
-    const beforeMention = inputValue.slice(0, mentionStartIndex);
-    const afterMention = inputValue.slice(
-      mentionStartIndex + mentionSearchQuery.length + 1,
-    );
-    const newValue = `${beforeMention}@${login} ${afterMention}`.trimEnd();
-
-    setInputValue(newValue);
-    setShowMentionAutocomplete(false);
-    setMentionSearchQuery("");
-    setMentionStartIndex(-1);
-    setInputKey((prev) => prev + 1);
-  };
-
-  const handleMentionCancel = () => {
-    setShowMentionAutocomplete(false);
-    setMentionSearchQuery("");
-    setMentionStartIndex(-1);
-  };
-
-  const handleSubmit = () => {
-    if (
-      inputValue.trim() &&
-      !isLoading &&
-      !disabled &&
-      !showMentionAutocomplete
-    ) {
-      onSubmit(inputValue);
-      setInputValue("");
-      setShowFullHelp(false);
-    }
-  };
-
-  const getDefaultHints = () => {
-    const hints: string[] = [];
-
-    if (availableCommands.length > 0) {
-      const nextCmd = availableCommands.find(
-        (cmd) => cmd.command.includes("next") || cmd.aliases?.includes("/next"),
-      );
-      if (nextCmd) {
-        hints.push(`${nextCmd.command} for next issue`);
-      }
-    }
-
-    if (availableCommands.length > 0) {
-      hints.push("? for help");
-    }
-
-    hints.push("ESC to go back");
-    hints.push("Ctrl + C to quit");
-    return hints;
-  };
-
-  const getAllCommandHints = () => {
-    if (availableCommands.length === 0) return [];
-
-    const hints: string[] = [];
-    let commandsToShow = availableCommands;
-
-    if (
-      inputValue.startsWith("/") &&
-      inputValue.length > 1 &&
-      !inputValue.includes(" ")
-    ) {
-      const searchTerm = inputValue.slice(1).toLowerCase();
-      commandsToShow = availableCommands.filter((cmd) => {
-        const commandName = cmd.command.split(" ")[0].slice(1).toLowerCase();
-        return commandName.startsWith(searchTerm);
-      });
-    }
-
-    commandsToShow.forEach((cmd) => {
-      hints.push(`${cmd.command} - ${cmd.description}`);
-    });
-
-    if (commandsToShow.length === 0 && inputValue.startsWith("/")) {
-      hints.push("No matching commands");
-    }
-
-    hints.push("? to hide help");
-    hints.push("ESC to go back");
-    return hints;
-  };
-
-  const defaultHints = getDefaultHints();
-  const allCommandHints = getAllCommandHints();
-
-  const renderedMessages = useMemo(() =>
-    messages.map((message) => renderMarkdown(message.content)),
-    [messages]
-  );
-
-  return (
-    <Box flexDirection="column">
-      {messages.length > 0 && (
-        <Box flexDirection="column" marginBottom={0}>
-          {messages.map((message, index) => (
-            <Box key={index} flexDirection="column" marginBottom={1}>
-              <Text bold color={message.role === "user" ? "cyan" : "yellow"}>
-                {message.role === "user" ? "You:" : "Baz:"}
-              </Text>
-              <Box paddingLeft={2}>
-                <Text>{renderedMessages[index]}</Text>
-              </Box>
-            </Box>
-          ))}
-        </Box>
-      )}
-
-      {isLoading && (
-        <Box marginBottom={1}>
-          <Text color="magenta">
-            <Spinner type="dots" />
-          </Text>
-          <Text color="magenta"> Thinking...</Text>
-        </Box>
-      )}
-
-      {!disabled && !isLoading && (
-        <Box flexDirection="column">
-          <Box
-            borderStyle="round"
-            borderColor="cyan"
-            paddingX={1}
-            width={terminalWidth}
-            flexShrink={1}
-          >
-            <TextInput
-              key={inputKey}
-              value={inputValue}
-              onChange={handleInputChange}
-              onSubmit={handleSubmit}
-              placeholder={placeholder}
-            />
+    return (
+      <Box flexDirection="column">
+        {messages.length > 0 && (
+          <Box flexDirection="column" marginBottom={0}>
+            {messages.map((message, index) => (
+              <MemoizedMessage key={index} message={message} />
+            ))}
           </Box>
-          {enableMentions &&
-            showMentionAutocomplete &&
-            reviewers.length > 0 && (
-              <MentionAutocomplete
-                reviewers={reviewers}
-                searchQuery={mentionSearchQuery}
-                onSelect={handleMentionSelect}
-                onCancel={handleMentionCancel}
-              />
-            )}
-          {!showMentionAutocomplete && (
-            <Box marginTop={1}>
-              <Text dimColor>
-                {showFullHelp
-                  ? allCommandHints.join("\n")
-                  : defaultHints.join("\n")}
-              </Text>
-            </Box>
-          )}
-        </Box>
-      )}
-    </Box>
-  );
-};
+        )}
+
+        {isLoading && (
+          <Box marginBottom={1}>
+            <Text color="magenta">
+              <Spinner type="dots" />
+            </Text>
+            <Text color="magenta"> Thinking...</Text>
+          </Box>
+        )}
+
+        {!disabled && !isLoading && (
+          <ChatInput
+            onSubmit={onSubmit}
+            placeholder={placeholder}
+            availableCommands={availableCommands}
+            enableMentions={enableMentions}
+            prId={prId}
+            onBack={onBack}
+            terminalWidth={terminalWidth}
+          />
+        )}
+      </Box>
+    );
+  },
+  (prevProps, nextProps) => {
+    // Don't re-render if only the messages array reference changed
+    // but the actual message contents are the same
+    if (prevProps.messages.length !== nextProps.messages.length) {
+      return false; // Re-render if message count changed
+    }
+
+    // Check if any message content actually changed
+    for (let i = 0; i < prevProps.messages.length; i++) {
+      if (
+        prevProps.messages[i].content !== nextProps.messages[i].content ||
+        prevProps.messages[i].role !== nextProps.messages[i].role
+      ) {
+        return false; // Re-render if message content changed
+      }
+    }
+
+    // Check other props
+    return (
+      prevProps.isLoading === nextProps.isLoading &&
+      prevProps.disabled === nextProps.disabled &&
+      prevProps.placeholder === nextProps.placeholder &&
+      prevProps.prId === nextProps.prId &&
+      prevProps.enableMentions === nextProps.enableMentions &&
+      prevProps.availableCommands?.length ===
+        nextProps.availableCommands?.length
+      // Intentionally skip onSubmit and onBack as they change on every render
+      // but don't affect the visual output when input isn't shown
+    );
+  },
+);
 
 export default ChatDisplay;
